requirements.txt
absl_py==2.3.1+computecanada
accelerate==1.10.1+computecanada
annotated_types==0.7.0+computecanada
antlr4_python3_runtime==4.9.3+computecanada
asttokens==3.0.0+computecanada
av==11.0.0+computecanada
braceexpand==0.1.7+computecanada
certifi==2025.8.3+computecanada
charset_normalizer==3.4.2+computecanada
click==8.2.1+computecanada
cloudpickle==3.1.1+computecanada
comm==0.2.3+computecanada
contourpy==1.3.1+computecanada
cycler==0.12.1+computecanada
debugpy==1.8.16+computecanada
decorator==5.2.1+computecanada
docker-pycreds==0.4.0+computecanada
docstring_parser==0.17.0+computecanada
einops==0.8.1+computecanada
executing==2.2.0+computecanada
filelock==3.18.0+computecanada
fonttools==4.59.0+computecanada
fsspec==2025.7.0+computecanada
ftfy==6.3.1+computecanada
gitdb==4.0.12+computecanada
gitpython==3.1.45+computecanada
grpcio==1.73.0+computecanada
h5py==3.13.0+computecanada
hf_xet==1.1.3+computecanada
huggingface-hub==0.34.3
hydra_core==1.3.2+computecanada
idna==3.10+computecanada
imageio==2.37.0+computecanada
importlib_metadata==8.7.0+computecanada
ipykernel==6.30.1+computecanada
ipython==9.4.0+computecanada
ipython_pygments_lexers==1.1.1+computecanada
jedi==0.19.2+computecanada
jinja2==3.1.6+computecanada
joblib==1.5.1+computecanada
jupyter_client==8.6.3+computecanada
jupyter_core==5.8.1+computecanada
kiwisolver==1.4.8+computecanada
lazy_loader==0.4+computecanada
lightning_utilities==0.15.2+computecanada
markdown==3.8.2+computecanada
markdown_it_py==3.0.0+computecanada
MarkupSafe==3.0.2+computecanada
matplotlib==3.10.0+computecanada
matplotlib_inline==0.1.7+computecanada
mdurl==0.1.2+computecanada
ml_collections==1.1.0+computecanada
monai==1.5.1
mpmath==1.3.0+computecanada
nest_asyncio==1.6.0+computecanada
networkx==3.5+computecanada
numpy==2.2.2+computecanada
omegaconf==2.3.0+computecanada
open_clip_torch==3.1.0
orjson==3.11.1+computecanada
packaging==25.0+computecanada
pandas==2.2.3+computecanada
parso==0.8.4+computecanada
pexpect==4.9.0+computecanada
pillow==11.1.0+computecanada
Pillow_SIMD==9.5.0.post2+computecanada
platformdirs==4.3.8+computecanada
peft==0.12.0
prompt_toolkit==3.0.51+computecanada
protobuf==5.29.5+computecanada
psutil==6.1.1+computecanada
ptyprocess==0.7.0+computecanada
pure_eval==0.2.3+computecanada
pydantic==2.11.7+computecanada
pydantic_core==2.33.2+computecanada
pydicom==3.0.1+computecanada
pygments==2.19.2+computecanada
pyparsing==3.2.3+computecanada
python_dateutil==2.9.0.post0+computecanada
python_dotenv==1.1.1+computecanada
pytz==2025.2+computecanada
pyvers==0.1.0+computecanada
PyYAML==6.0.2+computecanada
pyzmq==26.2.1+computecanada
qwen-vl-utils==0.0.11
regex==2024.11.6+computecanada
requests==2.32.4+computecanada
rich==14.1.0+computecanada
rich_argparse==1.7.1+computecanada
safetensors==0.5.3+computecanada
scikit_image==0.25.2+computecanada
scikit_learn==1.6.1+computecanada
scipy==1.15.1+computecanada
sentry_sdk==2.34.1+computecanada
setproctitle==1.3.4+computecanada
SimpleITK==2.3.1+computecanada
six==1.17.0+computecanada
smmap==5.0.2+computecanada
stack_data==0.6.3+computecanada
submitit==1.5.3
sympy==1.14.0+computecanada
tensorboard==2.20.0+computecanada
tensorboard_data_server==0.7.2+computecanada
tensordict==0.10.0+computecanada
termcolor==3.1.0+computecanada
threadpoolctl==3.6.0+computecanada
tifffile==2025.6.11+computecanada
timm==1.0.19+computecanada
tokenizers==0.21.1+computecanada
torch==2.7.1+computecanada
torchaudio==2.6.0+computecanada
torchmetrics==1.8.1+computecanada
torchvision==0.22.1+computecanada
tornado==6.5.1+computecanada
tqdm==4.67.1+computecanada
traitlets==5.14.3+computecanada
transformers==4.55.0+computecanada
triton==3.3.1+computecanada
typing_extensions==4.14.1+computecanada
typing_inspection==0.4.1+computecanada
tzdata==2025.2+computecanada
urllib3==2.5.0+computecanada
vit-pytorch==1.11.7
wandb==0.18.0+computecanada
wcwidth==0.2.13+computecanada
webdataset==1.0.2
werkzeug==3.1.3+computecanada
zipp==3.23.0+computecanada
